Move binary-specific configs to respective binary source files.
Previously, these configs were located in the common config.rs file in
the janus_server crate, which was unnecessary.
